Move system test to correct folder
